requirements.txt
aiohttp~=3.8.1
asyncpg~=0.26.0
python-dateutil~=2.8.2
pip~=22.2.2
attrs~=22.1.0
idna~=3.3
multidict~=6.0.2
yarl~=1.8.1
aiosignal~=1.2.0
frozenlist~=1.3.1
six~=1.16.0
setuptools~=65.3.0
httpx~=0.23.0